separate components, fix review
