+++ components/pages/game/components/constants.js
@@ -1,0 +1,1 @@
+export const LIMIT_QUESTIONS = 4;
