+++ components/pages/game/components/styled.ts
