+++ components/pages/game/index.ts
@@ -1,1 +1,1 @@
-export { default } from './GamePage';
+export { default } from './Game';
