+++ components/pages/index/HomePage.js
@@ -5,21 +5,20 @@ import { withApolloClient } from 'lib/withApolloClient';
 
 import DefaultTemplate from 'components/shared/templates/DefaultTemplate';
 import { NotifierProvider } from 'contexts/NotifierContext';
-
-import Button from 'components/shared/atoms/ButtonWithIcon';
+import Link from 'next/link';
+import ButtonWithIcon from 'components/shared/atoms/ButtonWithIcon';
 import { component as GoogleIcon } from 'public/images/icons/google-icon.svg';
 import useSignIn from 'lib/apollo/hooks/actions/useSignIn';
 import { GAME } from 'config/routes';
 import { signInWithGoogle } from 'lib/auth/signInWithGoogle';
 import { useCurrentUser } from 'lib/apollo/hooks/state/currentUser';
-import { useRouter } from 'next/router';
+import ButtonedLink from 'components/shared/atoms/ButtonedLink';
 import { Title, PageContent, Content, Oranization, TagLine, Description, PreviewImg, ImgGroup } from './styled';
 
 const googleClientId = process.env.GOOGLE_CLIENT_ID;
 const HomePage = () => {
   const { user } = useCurrentUser();
   const [signIn] = useSignIn();
-  const router = useRouter();
 
   const signInWithGoogleHandler = async () => {
     try {
@@ -32,9 +31,7 @@ const HomePage = () => {
       console.error(error);
     }
   };
-  const handleClick = () => {
-    router.push(GAME);
-  };
+
   return (
     <NotifierProvider>
       <DefaultTemplate>
@@ -47,11 +44,11 @@ const HomePage = () => {
               speed.
             </TagLine>
             {user ? (
-              <button type="button" style={{ width: '200px', height: '60px' }} onClick={handleClick}>
-                начать новую игру
-              </button>
+              <Link href={GAME} passHref>
+                <ButtonedLink>Начать новую игру</ButtonedLink>
+              </Link>
             ) : (
-              <Button onClick={signInWithGoogleHandler} icon={<GoogleIcon />} text="Login with Google" />
+              <ButtonWithIcon onClick={signInWithGoogleHandler} icon={<GoogleIcon />} text="Login with Google" />
             )}
             <Description>to save game results and have access to the leaderboard</Description>
           </Content>
