+++ interfaces/questionType.ts
@@ -1,0 +1,10 @@
+import { Colors } from 'public/styles/theme';
+
+export default interface IQuestion {
+  id: number;
+  name: string;
+  fullName: string;
+  wrongAnswers: string[];
+  answerStatusColor: Colors;
+  options: string[];
+}
